moved prepare_handle_data to utils
