+++ src/piddiplatsch/utils/models.py
@@ -1,5 +1,7 @@
+import json
 import logging
 from datetime import datetime
+from typing import Any
 from uuid import NAMESPACE_URL, uuid3
 
 from dateutil.parser import isoparse
@@ -16,6 +18,30 @@ def build_handle(pid: str, as_uri: bool = False) -> str:
     else:
         handle = f"{prefix}/{pid}"
     return handle
+
+
+def prepare_handle_data(record: dict[str, Any]) -> dict[str, str]:
+    """Prepare handle record fields: serialize list/dict values, skip None, convert datetime."""
+    prepared: dict[str, str] = {}
+
+    for key, value in record.items():
+        if value is None:
+            continue
+
+        if isinstance(value, list | dict):
+
+            def serialize(obj):
+                if isinstance(obj, datetime):
+                    return obj.isoformat()
+                return obj
+
+            value = json.dumps(value, default=serialize)
+        elif isinstance(value, datetime):
+            value = value.isoformat()
+
+        prepared[key] = value
+
+    return prepared
 
 
 def item_pid(item_id: str) -> str:
